+++ src/components/class/Chat.jsx
@@ -3,7 +3,7 @@ import styled from "styled-components";
 import UserImg from "../../elements/molecules/UserImg";
 import { Text } from "../../styles/Text";
 
-const Chat = ({ username, profile_img, message }) => {
+const Chat = ({ nickname: username, profile_img, message }) => {
   return (
     <StWrapper>
       <UserImg src={profile_img} alt={`profile_${username}`} />
