+++ src/components/class/ClassFrame.jsx
@@ -1,3 +1,7 @@
+import { useParams } from "react-router-dom";
+import { useDispatch } from "react-redux";
+import { useEffect } from "react";
+import { __getClassInfo } from "../../modules/redux/cookingClass";
 import styled from "styled-components";
 
 import GridTemplate from "../../elements/templates/GridTemplate";
@@ -5,6 +9,14 @@ import ChatFrame from "./ChatFrame";
 // import VideoFrame from "./VideoFrame";
 
 const ClassFrame = () => {
+  const { classId } = useParams();
+
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(__getClassInfo({ classId }));
+  }, [dispatch, classId]);
+
   return (
     <GridTemplate>
       <StClassSection>
