+++ src/components/classes/Class.jsx
@@ -1,8 +1,5 @@
-import { useDispatch } from "react-redux";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
-
-import { __getPrevChats } from "../../modules/redux/cookingClass";
 
 const Class = ({
   class_id,
@@ -11,13 +8,10 @@ const Class = ({
   class_img,
   viewer_nums,
 }) => {
-  const dispatch = useDispatch();
-
   const navigate = useNavigate();
 
   const clickHandler = () => {
-    dispatch(__getPrevChats({ class_id }));
-    navigate(`/class/${redis_class_id}/subscriber`);
+    navigate(`/class/${class_id}/${redis_class_id}`);
   };
 
   return (
